Revert "Fix broken api"
